Replicate HDAudio fixes to other drivers using same template
1) NonAlignedAddress needs to be assigned from address before address
gets aligned
2) Extend mask to full long size (64 or 32 bits) so that on 64-bit upper
part of address is not zeroed
3) Add checkme for freeing of aligned pointer, where non-aligned should
be freed.

+++ workbench/devs/AHI/Drivers/EMU10kx/linuxsupport.c
@@ -43,7 +43,7 @@ AllocPages( size_t size, ULONG req )
   if( address != NULL )
   {
     a = (unsigned long) address;
-    a = (a + PAGE_SIZE - 1 + sizeof(APTR)) & ~(PAGE_SIZE - 1); //(((unsigned long) (a + 4096)) / 4096) * 4096; // get a 4K-aligned memory pointer
+    a = (a + PAGE_SIZE - 1 + sizeof(APTR)) & ~((unsigned long)PAGE_SIZE - 1); //(((unsigned long) (a + 4096)) / 4096) * 4096; // get a 4K-aligned memory pointer
     ((APTR *)a)[-1] = address;
     address = (void *) a;
   }
